created separate folder and files for types
