+++ src/pages/login/login.tsx
@@ -1,7 +1,6 @@
 import { FC } from 'react';
-import { TAppProps } from '../../app';
 
-export const LoginPage: FC<TAppProps> = () => (
+export const LoginPage: FC = () => (
   <main className="page__main page__main--login">
     <div className="page__login-container container">
       <section className="login">
