+++ src/types/user.ts
@@ -1,0 +1,7 @@
+export type THost = {
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+  email: string;
+  token: string;
+};
